feat(ProductionOpenApiService, NetworkClient): API 호출을 통한 데이터를 받는 기능들과 데이터에 맞는 DTO 추가

+++ src/main/java/com/hokago_memories/domain/Tier.java
@@ -2,6 +2,6 @@ package com.hokago_memories.domain;
 
 public record Tier(
         String tierName,
-        int tierPoint
+        double tierPoint
 ) {
 }
